LPS-85052 Adapt SamlBinding to new OpenSAML message API
OpenSAML 3 message API has suffered considerable changes. We need to use
now a different interface and also the lifecycle of the components have
changed. They are no longer singletons that can be applied to several
messages, now it looks like they have to be setup and initialized per use,
so we are providing Suppliers for that.